+++ src/hybran/MCL.py
@@ -1,8 +1,11 @@
-import subprocess
 import logging
+import os
+import subprocess
 import sys
 import time
+
 from . import config
+
 
 def execute_mcxdeblast(blast):
     hybran_tmp_dir = config.hybran_tmp_dir
@@ -19,24 +22,6 @@ def execute_mcxdeblast(blast):
         text=True,
     )
     return mcxdeblast_out
-
-
-def execute_mcl():
-    hybran_tmp_dir = config.hybran_tmp_dir
-    mcl_cmd = ['mcl',
-               hybran_tmp_dir + '/mcxdeblast_results',
-               '--abc',
-               '-I', '1.5',
-               '-o', hybran_tmp_dir + '/mcl',
-               '-q', 'x',
-               '-V', 'all']
-    mcl_out = subprocess.run(
-        mcl_cmd,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT,
-        text=True,
-    )
-    return mcl_out
 
 
 def inflate_mcl_clusters(mcl_output, cdhit_groups, gene_key):
@@ -64,7 +49,7 @@ def writer(lines, output_name):
             out.write('\t'.join(line) + '\n')
 
 
-def run_mcl(in_blast, cdhit_clusters, out_name, gene_names):
+def run_mcl(in_blast, cdhit_clusters, out_name, gene_names, inflation):
     hybran_tmp_dir = config.hybran_tmp_dir    
     logger = logging.getLogger('MCL')
     logger.info('Running MCL')
@@ -76,15 +61,27 @@ def run_mcl(in_blast, cdhit_clusters, out_name, gene_names):
         logger.error('\n' + mcx_ps.stdout)
         logger.error('mcxdeblast failed.')
     time.sleep(5)
-    mcl_ps = execute_mcl()
+    mcl_cmd = [
+        'mcl',
+        os.path.join(hybran_tmp_dir, 'mcxdeblast_results'),
+        '--abc',
+        '-I', str(inflation),
+        '-o', os.path.join(hybran_tmp_dir, 'mcl'),
+        '-q', 'x',
+        '-V', 'all',
+    ]
+    mcl_ps = subprocess.run(
+        mcl_cmd,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.STDOUT,
+        text=True,
+    )
     try:
         mcl_ps.check_returncode()
-        logger.debug('mcl failed')
     except subprocess.CalledProcessError:
         logger.error('\n' + mcl_ps.stdout)
         logger.error('mcl failed')
     time.sleep(5)
-    logger.info('Re-inflating CDHIT clusters in MCL clusters')
     output = inflate_mcl_clusters(mcl_output=hybran_tmp_dir + '/mcl',
                                   cdhit_groups=cdhit_clusters,
                                   gene_key=gene_names)