+++ src/hybran/run.py
@@ -127,10 +127,13 @@ def clustering(all_genomes, target_genomes, nproc, seq_ident, seq_covg):
                 seq_covg=config.cnf.blast.min_coverage,
             )
         if 'clustered_proteins' not in os.listdir(os.getcwd()):
-            MCL.run_mcl(in_blast=hybran_tmp_dir + '/blast_results',
-                        cdhit_clusters=clusters,
-                        out_name='clustered_proteins',
-                        gene_names=gff_gene_dict)
+            MCL.run_mcl(
+                in_blast=os.path.join(hybran_tmp_dir, 'blast_results'),
+                cdhit_clusters=clusters,
+                out_name='clustered_proteins',
+                gene_names=gff_gene_dict,
+                inflation=config.cnf.mcl_inflation,
+            )
     os.chdir(c)
     parseClustering.parseClustersUpdateGBKs(target_gffs=all_genomes,
                                             clusters=hybran_tmp_dir +